showroom/usecase-010/requirements.txt
pywebview>=4.0
pillow>=9.0.0
